requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-groq
faiss-cpu
sentence-transformers
gitpython
streamlit
python-dotenv
slowapi
psycopg2-binary